Adds module for Enigma
