+++ test/test_helper.rb
@@ -12,3 +12,4 @@ require './lib/key'
 require './lib/offset'
 require './lib/shift'
 require './lib/enigma'
+require './modules/enigma_helper_module'
